Hook up Winch to Wasmtime
This commit adds an off-by-default feature to the `wasmtime` crate
called `winch` which will compile in support for Winch. This
additionally adds a new `Strategy::Winch` enumeration to select the
winch compiler.

This is intended to get some plumbing out of the way to get things
hooked up and eventually have a CLI option to choose Winch when it's
ready.

Additionally this commit refactors the `wasmtime-cranelift` compiler
builder implementation to be shared between Winch and Cranelift as a
proof-of-concept of sharing some Wasmtime boilerplate.