Fill out Winch's `append_code` function
This is the last major function required to get Winch hooked up into
Wasmtime. The `append_code` function is more-or-less a glorified
wrapper around the `MachTextSectionBuilder` type so most of the changes
here are moving things around so everything is accessible to both
Cranelift and Winch without all that much duplication.